Fix pmd error due to signing jars
jar signing plugin and pmd plugin don't work well together
with multi-module builds: we end up with partially signed
packages - one part is signed from the previous module,
but for the current module only target/classes is on the
auxclasspath which is not signed. If there are classes in the
same packages (such as test classes) this will fail with

SecurityException: ... signer information does not match
signer information of other classes in the same package

For now, we run the builds twice - first without signing
and then with signing (but without pmd).